--- src/POC/templates/index.html
@@ -1,13 +1,0 @@
-<html>
-
-<head>
-    <link rel="stylesheet" type="text/css" href="/static/css/index.css" />
-    <script type='text/javascript' src="{{url_for('static', filename='js/index.js')}}"></script>
-</head>
-
-<body>
-    <h1>Our Home Page</h1>
-    <p>Food is {{ food }}</p>
-</body>
-
-</html>
