+++ src/utrition-backend/settingslocal.py
@@ -1,0 +1,1 @@
+DEBUG = True
